Whiteboard/requirements.txt
Pillow
PyQt6